rename utils, update path, add styles
