+++ public/index.html
@@ -5,11 +5,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Pizza-PJ</title>
-    <style>
-      .price {
-        margin-top: 10px;
-      }
-    </style>
   </head>
   <body>
     <div id="header">
@@ -24,7 +19,7 @@
         <ul id="order-list">
           <span id="orderspan">Выберите ингредиенты</span>
         </ul>
-        <div class="price">Цена: <span id="price">23 p.</span></div>
+        <div id="price-block" class="price">Цена: <span id="price">23 p.</span></div>
       </div>
       <div id="order" class="order">
         <div id="ordersquare" class="order-square">
@@ -117,6 +112,5 @@
         </div>
       </div>
     </div>
-    <script src="script.js"></script>
   </body>
 </html>
