+++ src/index.js
@@ -5,3 +5,4 @@ import addClassesToOrder from '../src/order/index';
 import addClassesToIngredients from '../src/ingredients/index';
 import ingredients from '../src/ingredients'
 
+
