+++ src/ingredients/index.js
@@ -1,7 +1,8 @@
 import styles from "./ingredients.module.scss";
-import { addClassFoundById } from "../helpers/addClassFoundById";
-import { addClassFoundByClass} from '../helpers/addClassFoundByClass';
-/* import { addBackground } from '../helpers/addbackground'; */
+import { addClassFoundById } from "../utils/addClassFoundById";
+import { addClassFoundByClass} from '../utils/addClassFoundByClass';
+
+
 
 addClassFoundById("ingredients", styles.ingredients);
 addClassFoundById("ingredients", styles.container);
@@ -9,33 +10,7 @@ addClassFoundByClass("input-div", styles.inputdiv);
 addClassFoundByClass("input1", styles.input);
 addClassFoundByClass("checkbox-label", styles.checkboxlabel);
 
-class Pizza {
-    constructor() {
-      this.ingredients = {
-        dough: [],
-        meat: [],
-        vegetables: [],
-        sauce: []
-      };
-      this.price = 23;
-    }
-  
-    addIngredient(category, ingredient) {
-      this.ingredients[category].push(ingredient);
-    }
-  
-    removeIngredient(category, ingredient) {
-      const index = this.ingredients[category].indexOf(ingredient);
-      if (index > -1) {
-        this.ingredients[category].splice(index, 1);
-      }
-    }
-  
-    calculatePrice() {
-      // Calculate the price based on the selected ingredients
-      // Implementation logic goes here
-      // ...
-      return this.price;
-    }
-  }
+
+
+
