+++ src/order/order.module.scss
@@ -1,6 +1,5 @@
 @import "../variables.scss";
 
-
 .order {
   background-color: $pink;
   width: 50%;
@@ -66,10 +65,16 @@
     font-size: 18px;
     transition: transform 0.4s ease;
     cursor: pointer;
-
-    &:hover {
-      transform: scale(1.1);
-      color: $black;
-    }
   }
+  .enabled {
+    background-color: $green;
+    border: 1px solid $green;
+  }
+  .enabled:hover {
+    transform: scale(1.1);
+    color: $black;
+  } 
+
 }
+
+
