+++ src/orderList/index.js
@@ -1,6 +1,7 @@
 import styles from "./orderList.module.scss";
-import { addClassFoundById } from "../helpers/addClassFoundById";
+import { addClassFoundById } from "../utils/addClassFoundById";
 
 addClassFoundById("wrapper-main", styles.wrapper);
 addClassFoundById("wrapper-main", styles.container);
 addClassFoundById("viewingarea", styles.viewingarea);
+addClassFoundById("price-block", styles.price);
