+++ src/orderList/orderList.module.scss
@@ -31,8 +31,8 @@
   
       .price {
         align-self: flex-end;
-        padding-right: 50px;
         font-size: 21px;
+        padding: 0 50px 10px 0;
       }
     }
 }
