+++ src/utils/addClassFoundByClass.js
