+++ src/variables.scss
@@ -4,5 +4,6 @@ $yellow: rgb(238, 193, 79);
 $pink:  rgb(252, 188, 167);
 $red: rgb(206, 29, 29);
 $haki: darkkhaki;
+$green: rgb(5, 128, 5);
 
 $max-width: 1100px;
